// src/constants/error-type.ts
export enum ErrorType {
  None = 'None',
  SSO = 'SSO',
  NotAuthenticated = 'NotAuthenticated',
  Unknown = 'Unknown',
  SystemDown = 'SystemDown',
  BadRequest = 'BadRequest',
  Forbidden = 'Forbidden',
  NotFound = 'NotFound',
  MethodNotAllowed = 'MethodNotAllowed',
  NotAcceptable = 'NotAcceptable',
  Timeout = 'Timeout',
  Unavailable = 'Unavailable',
}